requirements.txt
python-telegram-bot==20.7
pdfkit==1.0.0
numpy>=1.21.0
scipy>=1.7.0
pytz>=2023.3
rasch-pkg>=0.1.0
openpyxl>=3.1.0

